Add defpclass and prevalence-object
And some minor unbound slot handling.
The reason for defpclass is convenience,
so we don't have to boilerplate-specify
:metaclass prevalence-class and
superclass prevalence-object.

The reason for prevalence-object is so
we have something to specialize on for
reinitialize-instance and other 'class'
specializing generic functions (as opposed
to things so far, which has specialized
on the metaclass). A slight nuisance.